Feature/flash call skeleton

See merge request client-sdk/sinchverification!10

+++ library/verification-sms/src/main/java/com/sinch/smsverification/initialization/SmsInitializationDetails.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class SmsInitializationDetails(
     @SerialName("template") val template: String,
-    @SerialName("interceptionTimeout") val interceptionTimeout: Int
+    @SerialName("interceptionTimeout") val interceptionTimeout: Long
 )
